AI Newsletter/requirements.txt
# AI / LLM  — FREE tier, no credit card (1,500 req/day)
google-generativeai>=0.7.0

# Scraping
feedparser>=6.0.11
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.0

# Config
pyyaml>=6.0.2

# Graph DB
neo4j>=5.20.0

# Email delivery
resend>=2.0.0

# Templating
jinja2>=3.1.4

# Local dev only
python-dotenv>=1.0.1
